(#4794) Refactoring - split SchedulerSubscriber to smaller class
