+++ app/bundles/ReportBundle/EventListener/SchedulerSubscriber.php
@@ -11,12 +11,10 @@
 
 namespace Mautic\ReportBundle\EventListener;
 
-use Mautic\CoreBundle\Form\DataTransformer\ArrayStringTransformer;
-use Mautic\EmailBundle\Helper\MailHelper;
 use Mautic\ReportBundle\Event\ReportScheduleSendEvent;
 use Mautic\ReportBundle\ReportEvents;
+use Mautic\ReportBundle\Scheduler\Model\SendSchedule;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-use Symfony\Component\Translation\TranslatorInterface;
 
 /**
  * Class SchedulerSubscriber.
@@ -24,19 +22,13 @@ use Symfony\Component\Translation\TranslatorInterface;
 class SchedulerSubscriber implements EventSubscriberInterface
 {
     /**
-     * @var MailHelper
+     * @var SendSchedule
      */
-    private $mailer;
+    private $sendSchedule;
 
-    /**
-     * @var TranslatorInterface
-     */
-    private $translator;
-
-    public function __construct(MailHelper $mailer, TranslatorInterface $translator)
+    public function __construct(SendSchedule $sendSchedule)
     {
-        $this->mailer     = $mailer->getMailer();
-        $this->translator = $translator;
+        $this->sendSchedule = $sendSchedule;
     }
 
     /**
@@ -51,27 +43,9 @@ class SchedulerSubscriber implements EventSubscriberInterface
 
     public function onScheduleSend(ReportScheduleSendEvent $event)
     {
-        $transformer = new ArrayStringTransformer();
-        $scheduler   = $event->getScheduler();
-        $report      = $scheduler->getReport();
-        $file        = $event->getFile();
-        $emails      = $transformer->reverseTransform($report->getToAddress());
-        $date        = new \DateTime();
-        $subject     = $this->translator->trans(
-            'mautic.report.schedule.email.subject',
-            ['%report_name%' => $report->getName(), '%date%' => $date->format('Y-m-d')]
-        );
-        $message = $this->translator->trans(
-            'mautic.report.schedule.email.message',
-            ['%report_name%' => $report->getName(), '%date%' => $date->format('Y-m-d')]
-        );
+        $scheduler = $event->getScheduler();
+        $file      = $event->getFile();
 
-        $this->mailer->setTo($emails);
-        $this->mailer->setSubject($subject);
-        $this->mailer->setBody($message);
-        $this->mailer->parsePlainText($message);
-        $this->mailer->attachFile($file);
-
-        $this->mailer->send(true);
+        $this->sendSchedule->send($scheduler, $file);
     }
 }
